Add compareAnswers to object SQLTestUtils
